src/lewis/requirements.txt
 django==2.2.14 
 django-cors-headers==3.11.0 
 django-environ==0.4.5 
 djangorestframework==3.13.1 
 djongo==1.3.4 
 dnspython==2.2.0 
 pyjwt==2.3.0 
 pymongo==3.12.0
